+++ src/core/src/main/scala/com/daromi/cascade/core/math/Series.scala
@@ -1,12 +1,12 @@
 package com.daromi.cascade.core.math
 
-def triangular(n: Long): Long =
-  require(n >= 1L, "n must be positive")
+def triangular(n: Int): Int =
+  require(n >= 1, "n must be positive")
 
-  n * (n + 1L) / 2L
+  n * (n + 1) / 2
 
-def triangular1(n: Long): Long =
-  require(n >= 1L, "n must be positive")
+def triangular1(n: Int): Int =
+  require(n >= 1, "n must be positive")
 
-  if (n == 1L) 1L
-  else n + triangular(n - 1L)
+  if (n == 1) 1
+  else n + triangular(n - 1)
